IGVF-971-file-dl-memory
* Disable the file-download button if API_URL has no value.
* Have all file pages detect and redirect download links to the NextJS server. Rename these files so we can get the slub after the path on file pages.
* Remove a stray console.log.
* Add tests for the new functions in files.ts.
