+++ pages/reference-files/[...id].js
@@ -1,6 +1,6 @@
 // node_modules
-import PropTypes from "prop-types";
 import Link from "next/link";
+import PropTypes from "prop-types";
 // components
 import AlternateAccessions from "../../components/alternate-accessions";
 import Attribution from "../../components/attribution";
@@ -20,7 +20,6 @@ import { FileHeaderDownload } from "../../components/file-download";
 import JsonDisplay from "../../components/json-display";
 import ObjectPageHeader from "../../components/object-page-header";
 import PagePreamble from "../../components/page-preamble";
-import SeparatedList from "../../components/separated-list";
 // lib
 import buildAttribution from "../../lib/attribution";
 import buildBreadcrumbs from "../../lib/breadcrumbs";
@@ -31,73 +30,75 @@ import {
 } from "../../lib/common-requests";
 import errorObjectToProps from "../../lib/errors";
 import FetchRequest from "../../lib/fetch-request";
+import {
+  checkForFileDownloadPath,
+  convertFileDownloadPathToFilePagePath,
+} from "../../lib/files";
 import { isJsonFormat } from "../../lib/query-utils";
 
-export default function AlignmentFile({
-  attribution,
-  alignmentFile,
-  fileSet = null,
+export default function ReferenceFile({
+  referenceFile,
+  fileSet,
   documents,
   derivedFrom,
   derivedFromFileSets,
   fileFormatSpecifications,
-  referenceFiles,
+  attribution = null,
   isJson,
 }) {
   return (
     <>
       <Breadcrumbs />
-      <EditableItem item={alignmentFile}>
+      <EditableItem item={referenceFile}>
         <PagePreamble>
           <AlternateAccessions
-            alternateAccessions={alignmentFile.alternate_accessions}
+            alternateAccessions={referenceFile.alternate_accessions}
           />
         </PagePreamble>
-        <ObjectPageHeader item={alignmentFile} isJsonFormat={isJson}>
-          <FileHeaderDownload file={alignmentFile} />
+        <ObjectPageHeader item={referenceFile} isJsonFormat={isJson}>
+          <FileHeaderDownload file={referenceFile} />
         </ObjectPageHeader>
-        <JsonDisplay item={alignmentFile} isJsonFormat={isJson}>
+        <JsonDisplay item={referenceFile} isJsonFormat={isJson}>
           <DataPanel>
             <DataArea>
               <FileDataItems
-                item={alignmentFile}
+                item={referenceFile}
                 fileSet={fileSet}
               ></FileDataItems>
             </DataArea>
           </DataPanel>
-          <DataAreaTitle>Alignment Details</DataAreaTitle>
-          <DataPanel>
-            <DataArea>
-              {referenceFiles.length > 0 && (
-                <>
-                  <DataItemLabel>Reference Files</DataItemLabel>
-                  <DataItemValue>
-                    <SeparatedList>
-                      {referenceFiles.map((file) => (
-                        <Link href={file["@id"]} key={file["@id"]}>
-                          {file.accession}
+          {(referenceFile.assembly || referenceFile.source_url) && (
+            <>
+              <DataAreaTitle>Reference Details</DataAreaTitle>
+              <DataPanel>
+                <DataArea>
+                  {referenceFile.assembly && (
+                    <>
+                      <DataItemLabel>Genome Assembly</DataItemLabel>
+                      <DataItemValue>{referenceFile.assembly}</DataItemValue>
+                    </>
+                  )}
+                  {referenceFile.source_url && (
+                    <>
+                      <DataItemLabel>Source URL</DataItemLabel>
+                      <DataItemValue>
+                        <Link
+                          href={referenceFile.source_url}
+                          key={referenceFile.source_url}
+                        >
+                          {referenceFile.source_url}
                         </Link>
-                      ))}
-                    </SeparatedList>
-                  </DataItemValue>
-                </>
-              )}
-              <>
-                <DataItemLabel>Redacted</DataItemLabel>
-                <DataItemValue>
-                  {alignmentFile.redacted ? "Yes" : "No"}
-                </DataItemValue>
-              </>
-              <DataItemLabel>Filtered</DataItemLabel>
-              <DataItemValue>
-                {alignmentFile.filtered ? "Yes" : "No"}
-              </DataItemValue>
-            </DataArea>
-          </DataPanel>
+                      </DataItemValue>
+                    </>
+                  )}
+                </DataArea>
+              </DataPanel>
+            </>
+          )}
           {derivedFrom.length > 0 && (
             <>
               <DataAreaTitle>
-                Files {alignmentFile.accession} Derives From
+                Files {referenceFile.accession} Derives From
               </DataAreaTitle>
               <DerivedFromTable
                 derivedFrom={derivedFrom}
@@ -124,40 +125,48 @@ export default function AlignmentFile({
   );
 }
 
-AlignmentFile.propTypes = {
-  // AlignmentFile object to display
-  alignmentFile: PropTypes.object.isRequired,
+ReferenceFile.propTypes = {
+  // ReferenceFile object to display
+  referenceFile: PropTypes.object.isRequired,
   // File set that contains this file
   fileSet: PropTypes.object,
   // Documents set associate with this file
-  documents: PropTypes.array.isRequired,
+  documents: PropTypes.array,
   // The file is derived from
-  derivedFrom: PropTypes.array.isRequired,
+  derivedFrom: PropTypes.array,
   // Filesets derived from files belong to
   derivedFromFileSets: PropTypes.arrayOf(PropTypes.object).isRequired,
   // Set of documents for file specifications
   fileFormatSpecifications: PropTypes.array.isRequired,
-  // Attribution for this file
-  attribution: PropTypes.object.isRequired,
-  // The file is derived from
-  referenceFiles: PropTypes.array.isRequired,
+  // Attribution for this ReferenceFile
+  attribution: PropTypes.object,
   // Is the format JSON?
   isJson: PropTypes.bool.isRequired,
 };
 
-export async function getServerSideProps({ params, req, query }) {
+export async function getServerSideProps({ params, req, query, resolvedUrl }) {
+  // Redirect to the file page if the URL is a file download link.
+  if (checkForFileDownloadPath(resolvedUrl)) {
+    return {
+      redirect: {
+        destination: convertFileDownloadPathToFilePagePath(resolvedUrl),
+        permanent: false,
+      },
+    };
+  }
+
   const isJson = isJsonFormat(query);
   const request = new FetchRequest({ cookie: req.headers.cookie });
-  const alignmentFile = await request.getObject(
-    `/alignment-files/${params.id}/`
+  const referenceFile = await request.getObject(
+    `/reference-files/${params.id}/`
   );
-  if (FetchRequest.isResponseSuccess(alignmentFile)) {
-    const fileSet = await request.getObject(alignmentFile.file_set, null);
-    const documents = alignmentFile.documents
-      ? await requestDocuments(alignmentFile.documents, request)
+  if (FetchRequest.isResponseSuccess(referenceFile)) {
+    const fileSet = await request.getObject(referenceFile.file_set, null);
+    const documents = referenceFile.documents
+      ? await requestDocuments(referenceFile.documents, request)
       : [];
-    const derivedFrom = alignmentFile.derived_from
-      ? await requestFiles(alignmentFile.derived_from, request)
+    const derivedFrom = referenceFile.derived_from
+      ? await requestFiles(referenceFile.derived_from, request)
       : [];
     const derivedFromFileSetPaths = derivedFrom
       .map((file) => file.file_set)
@@ -167,39 +176,35 @@ export async function getServerSideProps({ params, req, query }) {
       uniqueDerivedFromFileSetPaths.length > 0
         ? await requestFileSets(uniqueDerivedFromFileSetPaths, request)
         : [];
-    const fileFormatSpecifications = alignmentFile.file_format_specifications
+    const fileFormatSpecifications = referenceFile.file_format_specifications
       ? await requestDocuments(
-          alignmentFile.file_format_specifications,
+          referenceFile.file_format_specifications,
           request
         )
       : [];
-    const referenceFiles = alignmentFile.reference_files
-      ? await requestFiles(alignmentFile.reference_files, request)
-      : [];
     const breadcrumbs = await buildBreadcrumbs(
-      alignmentFile,
+      referenceFile,
       "accession",
       req.headers.cookie
     );
     const attribution = await buildAttribution(
-      alignmentFile,
+      referenceFile,
       req.headers.cookie
     );
     return {
       props: {
-        alignmentFile,
+        referenceFile,
         fileSet,
         documents,
         derivedFrom,
         derivedFromFileSets,
         fileFormatSpecifications,
-        pageContext: { title: alignmentFile.accession },
+        pageContext: { title: referenceFile.accession },
         breadcrumbs,
         attribution,
-        referenceFiles,
         isJson,
       },
     };
   }
-  return errorObjectToProps(alignmentFile);
+  return errorObjectToProps(referenceFile);
 }
