+++ pages/signal-files/[...id].js
@@ -1,6 +1,6 @@
 // node_modules
-import Link from "next/link";
 import PropTypes from "prop-types";
+import Link from "next/link";
 // components
 import AlternateAccessions from "../../components/alternate-accessions";
 import Attribution from "../../components/attribution";
@@ -20,6 +20,7 @@ import { FileHeaderDownload } from "../../components/file-download";
 import JsonDisplay from "../../components/json-display";
 import ObjectPageHeader from "../../components/object-page-header";
 import PagePreamble from "../../components/page-preamble";
+import SeparatedList from "../../components/separated-list";
 // lib
 import buildAttribution from "../../lib/attribution";
 import buildBreadcrumbs from "../../lib/breadcrumbs";
@@ -30,94 +31,86 @@ import {
 } from "../../lib/common-requests";
 import errorObjectToProps from "../../lib/errors";
 import FetchRequest from "../../lib/fetch-request";
-import { truthyOrZero } from "../../lib/general";
+import {
+  checkForFileDownloadPath,
+  convertFileDownloadPathToFilePagePath,
+} from "../../lib/files";
 import { isJsonFormat } from "../../lib/query-utils";
 
-export default function SequenceFile({
-  sequenceFile,
-  fileSet,
+export default function SignalFile({
+  attribution,
+  signalFile,
+  fileSet = null,
   documents,
   derivedFrom,
   derivedFromFileSets,
   fileFormatSpecifications,
-  attribution = null,
+  referenceFiles,
   isJson,
-  seqSpec = null,
-  sequencingPlatform = null,
 }) {
   return (
     <>
       <Breadcrumbs />
-      <EditableItem item={sequenceFile}>
+      <EditableItem item={signalFile}>
         <PagePreamble>
           <AlternateAccessions
-            alternateAccessions={sequenceFile.alternate_accessions}
+            alternateAccessions={signalFile.alternate_accessions}
           />
         </PagePreamble>
-        <ObjectPageHeader item={sequenceFile} isJsonFormat={isJson}>
-          <FileHeaderDownload file={sequenceFile} />
+        <ObjectPageHeader item={signalFile} isJsonFormat={isJson}>
+          <FileHeaderDownload file={signalFile} />
         </ObjectPageHeader>
-        <JsonDisplay item={sequenceFile} isJsonFormat={isJson}>
+        <JsonDisplay item={signalFile} isJsonFormat={isJson}>
           <DataPanel>
             <DataArea>
               <FileDataItems
-                item={sequenceFile}
+                item={signalFile}
                 fileSet={fileSet}
               ></FileDataItems>
             </DataArea>
           </DataPanel>
-          <DataAreaTitle>Sequencing Details</DataAreaTitle>
+          <DataAreaTitle>Signal Details</DataAreaTitle>
           <DataPanel>
             <DataArea>
-              {sequenceFile.flowcell_id && (
+              {referenceFiles.length > 0 && (
                 <>
-                  <DataItemLabel>Flowcell ID</DataItemLabel>
-                  <DataItemValue>{sequenceFile.flowcell_id}</DataItemValue>
-                </>
-              )}
-              {sequenceFile.illumina_read_type && (
-                <>
-                  <DataItemLabel>Illumina Read Type</DataItemLabel>
+                  <DataItemLabel>Reference Files</DataItemLabel>
                   <DataItemValue>
-                    {sequenceFile.illumina_read_type}
+                    <SeparatedList>
+                      {referenceFiles.map((file) => (
+                        <Link href={file["@id"]} key={file["@id"]}>
+                          {file.accession}
+                        </Link>
+                      ))}
+                    </SeparatedList>
                   </DataItemValue>
                 </>
               )}
-              {truthyOrZero(sequenceFile.read_count) && (
+              <>
+                <DataItemLabel>Strand Specificity</DataItemLabel>
+                <DataItemValue>{signalFile.strand_specificity}</DataItemValue>
+              </>
+              {"filtered" in signalFile && (
                 <>
-                  <DataItemLabel>Read Count</DataItemLabel>
-                  <DataItemValue>{sequenceFile.read_count}</DataItemValue>
-                </>
-              )}
-              {truthyOrZero(sequenceFile.mean_read_length) && (
-                <>
-                  <DataItemLabel>Read Length</DataItemLabel>
-                  <DataItemValue>{sequenceFile.mean_read_length}</DataItemValue>
-                </>
-              )}
-              {sequencingPlatform && (
-                <>
-                  <DataItemLabel>Sequencing Platform</DataItemLabel>
+                  <DataItemLabel>Filtered</DataItemLabel>
                   <DataItemValue>
-                    <Link href={sequencingPlatform["@id"]}>
-                      {sequencingPlatform.term_name}
-                    </Link>
+                    {signalFile.filtered ? "Yes" : "No"}
                   </DataItemValue>
                 </>
               )}
-              <DataItemLabel>Sequencing Run</DataItemLabel>
-              <DataItemValue>{sequenceFile.sequencing_run}</DataItemValue>
-              {sequenceFile.lane && (
+              {"normalized" in signalFile && (
                 <>
-                  <DataItemLabel>Lane</DataItemLabel>
-                  <DataItemValue>{sequenceFile.lane}</DataItemValue>
+                  <DataItemLabel>Normalized</DataItemLabel>
+                  <DataItemValue>
+                    {signalFile.normalized ? "Yes" : "No"}
+                  </DataItemValue>
                 </>
               )}
-              {seqSpec && (
+              {signalFile.start_view_position && (
                 <>
-                  <DataItemLabel>Associated seqspec File</DataItemLabel>
+                  <DataItemLabel>Start View Position</DataItemLabel>
                   <DataItemValue>
-                    <Link href={seqSpec["@id"]}>{seqSpec.accession}</Link>
+                    {signalFile.start_view_position}
                   </DataItemValue>
                 </>
               )}
@@ -126,7 +119,7 @@ export default function SequenceFile({
           {derivedFrom.length > 0 && (
             <>
               <DataAreaTitle>
-                Files {sequenceFile.accession} Derives From
+                Files {signalFile.accession} Derives From
               </DataAreaTitle>
               <DerivedFromTable
                 derivedFrom={derivedFrom}
@@ -153,40 +146,49 @@ export default function SequenceFile({
   );
 }
 
-SequenceFile.propTypes = {
-  // SequenceFile object to display
-  sequenceFile: PropTypes.object.isRequired,
+SignalFile.propTypes = {
+  // SignalFile object to display
+  signalFile: PropTypes.object.isRequired,
   // File set that contains this file
   fileSet: PropTypes.object,
   // Documents set associate with this file
-  documents: PropTypes.array,
+  documents: PropTypes.array.isRequired,
   // The file is derived from
-  derivedFrom: PropTypes.array,
+  derivedFrom: PropTypes.array.isRequired,
   // Filesets derived from files belong to
   derivedFromFileSets: PropTypes.arrayOf(PropTypes.object).isRequired,
   // Set of documents for file specifications
   fileFormatSpecifications: PropTypes.array.isRequired,
   // Attribution for this file
-  attribution: PropTypes.object,
+  attribution: PropTypes.object.isRequired,
+  // The file is derived from
+  referenceFiles: PropTypes.array.isRequired,
   // Is the format JSON?
   isJson: PropTypes.bool.isRequired,
-  // Linked seqspec configuration file
-  seqSpec: PropTypes.object,
-  // Sequencing platform ontology term object
-  sequencingPlatform: PropTypes.object,
 };
 
-export async function getServerSideProps({ params, req, query }) {
+export async function getServerSideProps({ params, req, query, resolvedUrl }) {
+  // Redirect to the file page if the URL is a file download link.
+  const isPathForFileDownload = checkForFileDownloadPath(resolvedUrl);
+  if (isPathForFileDownload) {
+    return {
+      redirect: {
+        destination: convertFileDownloadPathToFilePagePath(resolvedUrl),
+        permanent: false,
+      },
+    };
+  }
+
   const isJson = isJsonFormat(query);
   const request = new FetchRequest({ cookie: req.headers.cookie });
-  const sequenceFile = await request.getObject(`/sequence-files/${params.id}/`);
-  if (FetchRequest.isResponseSuccess(sequenceFile)) {
-    const fileSet = await request.getObject(sequenceFile.file_set, null);
-    const documents = sequenceFile.documents
-      ? await requestDocuments(sequenceFile.documents, request)
+  const signalFile = await request.getObject(`/signal-files/${params.id}/`);
+  if (FetchRequest.isResponseSuccess(signalFile)) {
+    const fileSet = await request.getObject(signalFile.file_set, null);
+    const documents = signalFile.documents
+      ? await requestDocuments(signalFile.documents, request)
       : [];
-    const derivedFrom = sequenceFile.derived_from
-      ? await requestFiles(sequenceFile.derived_from, request)
+    const derivedFrom = signalFile.derived_from
+      ? await requestFiles(signalFile.derived_from, request)
       : [];
     const derivedFromFileSetPaths = derivedFrom
       .map((file) => file.file_set)
@@ -196,40 +198,33 @@ export async function getServerSideProps({ params, req, query }) {
       uniqueDerivedFromFileSetPaths.length > 0
         ? await requestFileSets(uniqueDerivedFromFileSetPaths, request)
         : [];
-    const fileFormatSpecifications = sequenceFile.file_format_specifications
-      ? await requestDocuments(sequenceFile.file_format_specifications, request)
+    const fileFormatSpecifications = signalFile.file_format_specifications
+      ? await requestDocuments(signalFile.file_format_specifications, request)
       : [];
-    const seqSpec = sequenceFile.seqspec
-      ? await request.getObject(sequenceFile.seqspec, null)
-      : null;
-    const sequencingPlatform = sequenceFile.sequencing_platform
-      ? await request.getObject(sequenceFile.sequencing_platform, null)
-      : null;
+    const referenceFiles = signalFile.reference_files
+      ? await requestFiles(signalFile.reference_files, request)
+      : [];
     const breadcrumbs = await buildBreadcrumbs(
-      sequenceFile,
+      signalFile,
       "accession",
       req.headers.cookie
     );
-    const attribution = await buildAttribution(
-      sequenceFile,
-      req.headers.cookie
-    );
+    const attribution = await buildAttribution(signalFile, req.headers.cookie);
     return {
       props: {
-        sequenceFile,
+        signalFile,
         fileSet,
         documents,
         derivedFrom,
         derivedFromFileSets,
         fileFormatSpecifications,
-        pageContext: { title: sequenceFile.accession },
+        pageContext: { title: signalFile.accession },
         breadcrumbs,
         attribution,
+        referenceFiles,
         isJson,
-        seqSpec,
-        sequencingPlatform,
       },
     };
   }
-  return errorObjectToProps(sequenceFile);
+  return errorObjectToProps(signalFile);
 }
